+++ feature_base/src/main/kotlin/com/igorwojda/showcase/base/presentation/compose/composable/Lottie.kt
@@ -1,4 +1,4 @@
-package com.igorwojda.showcase.base.presentation.compose
+package com.igorwojda.showcase.base.presentation.compose.composable
 
 import androidx.annotation.RawRes
 import androidx.annotation.StringRes
